Add missing includes, func decls, remove unused locals
There's a lot of `(s32)Rng_Rand16()` in the project, maybe should look
into changing return type of it.
(some of the Rng_Rand16 calls that don't need `(s32)` could be using
implicit s32 ref, if `rng.h` wasn't included?)

+++ include/bodyprog/bodyprog.h
@@ -2926,7 +2926,7 @@ void GameFs_UniqueItemModelLoad(u8 itemId);
 
 void GameFs_MapItemsTextureLoad(s32 mapId);
 
-void func_800546A8(s32 arg0);
+void func_800546A8(u8 arg0);
 
 void func_8005487C(s32);
